+++ src/main/scala/org/podval/xml/Html.scala
@@ -4,12 +4,9 @@ import zio.blocks.chunk.Chunk
 import zio.blocks.html.Dom as XML
 
 // XML AST for ZIO Blocks HTML. Not a package given: `import Html.given` to
-// parse/write this AST; `import Html.toHtml` converts any `XmlAst` element.
+// parse/write this AST; convert with `element.to[Html.Element]`.
 object Html extends XmlAst[XML.Element]:
   given Html.type = this
-
-  extension [E: XmlAst](element: E)
-    def toHtml: Element = element.to[Element](using Html)
 
   override type Node = XML